+++ src/browserContentProvider.ts
@@ -1,9 +1,9 @@
 "use strict";
 import * as vscode from "vscode";
-import { Utility } from "./Utility";
+import { Utility } from "./utility";
 
 export class BrowserContentProvider implements vscode.TextDocumentContentProvider {
-    // private _onDidChange = new vscode.EventEmitter<vscode.Uri>();
+    private _onDidChange = new vscode.EventEmitter<vscode.Uri>();
 
     public provideTextDocumentContent() {
         const editor = vscode.window.activeTextEditor;
@@ -18,13 +18,12 @@ export class BrowserContentProvider implements vscode.TextDocumentContentProvide
 
         return `<iframe src="${uri}" frameBorder="0" width="100%" height="1000px" />`;
     }
-    /*
+
     get onDidChange(): vscode.Event<vscode.Uri> {
         return this._onDidChange.event;
     }
-    */
+
     public update(uri: vscode.Uri) {
-        console.log(uri);
-        // this._onDidChange.fire(uri);
+        this._onDidChange.fire(uri);
     }
 }