+++ src/extension.ts
@@ -1,29 +1,36 @@
 import * as vscode from "vscode";
 import { BrowserContentProvider } from "./browserContentProvider";
-import { Utility } from "./Utility";
+import { Server } from "./server";
+import { Utility } from "./utility";
 
 export function activate(context: vscode.ExtensionContext) {
-    Utility.startWebServer();
+    // start web server
+    startServer();
 
+    // provider settings.
     const provider = new BrowserContentProvider();
     const registration = vscode.workspace.registerTextDocumentContentProvider("http", provider);
-    let previewUri = vscode.Uri.parse("http://localhost");
-
-    /*
     vscode.workspace.onDidChangeTextDocument((e: vscode.TextDocumentChangeEvent) => {
         if (e.document === vscode.window.activeTextEditor.document) {
+            const previewUri = Utility.getUriOfActiveEditor();
             provider.update(previewUri);
         }
     });
-    */
 
+    // When configuration is changed, resume web server.
     vscode.workspace.onDidChangeConfiguration(() => {
-        Utility.resumeWebServer();
+        resumeServer();
         vscode.window.showInformationMessage("Resume the Web Server.");
     });
 
+    // When file is saved, reload browser.
+    vscode.workspace.onDidSaveTextDocument((e) => {
+        Server.reload(e.fileName);
+    });
+
     let disposable: any = vscode.commands.registerCommand("extension.preview", () => {
-        Utility.startWebServer();
+        const previewUri = Utility.getUriOfActiveEditor();
+
         // set ViewColumn
         let viewColumn: vscode.ViewColumn;
         if (vscode.window.activeTextEditor.viewColumn < 3) {
@@ -39,17 +46,35 @@ export function activate(context: vscode.ExtensionContext) {
     });
 
     let disposable2: any = vscode.commands.registerCommand("extension.launch", () => {
-        Utility.startWebServer();
         const uri = Utility.getUriOfActiveEditor();
         return vscode.commands.executeCommand("vscode.open", uri);
     });
 
     let disposable3: any = vscode.commands.registerCommand("extension.stop", () => {
-        Utility.stopWebServer();
+        Server.stop();
         vscode.window.showInformationMessage("Stop the Web Server successfully.");
     });
 
-    context.subscriptions.push(disposable, disposable2, disposable3, registration);
+    let disposable4: any = vscode.commands.registerCommand("extension.resume", () => {
+        resumeServer();
+        vscode.window.showInformationMessage("Resume the Web Server.");
+    });
+
+    context.subscriptions.push(disposable, disposable2, disposable3, disposable4, registration);
+}
+
+function startServer() {
+    const options = vscode.workspace.getConfiguration("previewServer");
+    const port = options.get("port") as number;
+    const isSync = options.get("sync") as boolean;
+    const rootPath = vscode.workspace.rootPath || vscode.window.activeTextEditor.document.fileName;
+
+    Server.start(rootPath, port, isSync);
+}
+
+function resumeServer() {
+    Server.stop();
+    startServer();
 }
 
 export function deactivate() {